[level 0] Title: 배열의 평균값, Time: 0.04 ms, Memory: 74.7 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.java
@@ -1,11 +1,11 @@
 class Solution {
     public double solution(int[] numbers) {
-        float res = 0.0f;
+        double res = 0.0f;
         int sum = 0;
         for (int i: numbers) {
             sum += i;
         }
-        res = (float) sum / numbers.length;
+        res = (double) sum / numbers.length;
         return res;
     }
 }
